[refactor] aarch64 el2 support (#31)
