+++ modules/axhal/src/platform/aarch64_common/boot.rs
@@ -1,8 +1,5 @@
 use aarch64_cpu::{asm, asm::barrier, registers::*};
 use core::ptr::addr_of_mut;
-use memory_addr::PhysAddr;
-//Todo: remove this, when hv is enabled, `MemAttr` is not used.
-#[cfg_attr(feature = "hv", allow(unused_imports))]
 use page_table_entry::aarch64::{MemAttr, A64PTE};
 use tock_registers::interfaces::{ReadWriteable, Readable, Writeable};
 
@@ -17,7 +14,6 @@ static mut BOOT_PT_L0: [A64PTE; 512] = [A64PTE::empty(); 512];
 #[link_section = ".data.boot_page_table"]
 static mut BOOT_PT_L1: [A64PTE; 512] = [A64PTE::empty(); 512];
 
-#[cfg(not(feature = "hv"))]
 unsafe fn switch_to_el1() {
     SPSel.write(SPSel::SP::ELx);
     SP_EL0.set(0);
@@ -61,7 +57,6 @@ unsafe fn switch_to_el1() {
     }
 }
 
-#[cfg(not(feature = "hv"))]
 unsafe fn init_mmu() {
     MAIR_EL1.set(MemAttr::MAIR_VALUE);
 
@@ -106,7 +101,6 @@ unsafe fn init_boot_page_table() {
 }
 
 /// The earliest entry point for the primary CPU.
-#[cfg(not(feature = "hv"))]
 #[naked]
 #[no_mangle]
 #[link_section = ".text.boot"]
@@ -148,7 +142,6 @@ unsafe extern "C" fn _start() -> ! {
 }
 
 /// The earliest entry point for the secondary CPUs.
-#[cfg(not(feature = "hv"))]
 #[cfg(feature = "smp")]
 #[naked]
 #[no_mangle]
@@ -177,148 +170,4 @@ unsafe extern "C" fn _start_secondary() -> ! {
         entry = sym crate::platform::rust_entry_secondary,
         options(noreturn),
     )
-}
-
-#[cfg(feature = "hv")]
-unsafe fn switch_to_el2() {
-    SPSel.write(SPSel::SP::ELx);
-    let current_el = CurrentEL.read(CurrentEL::EL);
-
-    if current_el == 3 {
-        SCR_EL3.write(
-            SCR_EL3::NS::NonSecure + SCR_EL3::HCE::HvcEnabled + SCR_EL3::RW::NextELIsAarch64,
-        );
-        SPSR_EL3.write(
-            SPSR_EL3::M::EL2h
-                + SPSR_EL3::D::Masked
-                + SPSR_EL3::A::Masked
-                + SPSR_EL3::I::Masked
-                + SPSR_EL3::F::Masked,
-        );
-        ELR_EL3.set(LR.get());
-        SP_EL1.set(BOOT_STACK.as_ptr_range().end as u64);
-        // This should be SP_EL2. To
-        asm::eret();
-    }
-}
-
-#[cfg(feature = "hv")]
-unsafe fn init_mmu_el2() {
-    // Set EL1 to 64bit.
-    HCR_EL2.write(HCR_EL2::RW::EL1IsAarch64);
-
-    // Device-nGnRE memory
-    let attr0 = MAIR_EL2::Attr0_Device::nonGathering_nonReordering_EarlyWriteAck;
-    // Normal memory
-    let attr1 = MAIR_EL2::Attr1_Normal_Inner::WriteBack_NonTransient_ReadWriteAlloc
-        + MAIR_EL2::Attr1_Normal_Outer::WriteBack_NonTransient_ReadWriteAlloc;
-    MAIR_EL2.write(attr0 + attr1); // 0xff_04
-
-    // Enable TTBR0 and TTBR1 walks, page size = 4K, vaddr size = 48 bits, paddr size = 40 bits.
-    let tcr_flags0 = TCR_EL2::TG0::KiB_4
-        + TCR_EL2::SH0::Inner
-        + TCR_EL2::ORGN0::WriteBack_ReadAlloc_WriteAlloc_Cacheable
-        + TCR_EL2::IRGN0::WriteBack_ReadAlloc_WriteAlloc_Cacheable
-        + TCR_EL2::T0SZ.val(16);
-    TCR_EL2.write(TCR_EL2::PS::Bits_40 + tcr_flags0);
-    barrier::isb(barrier::SY);
-
-    let root_paddr = PhysAddr::from(BOOT_PT_L0.as_ptr() as usize).as_usize() as _;
-    TTBR0_EL2.set(root_paddr);
-
-    // Flush the entire TLB
-    crate::arch::flush_tlb(None);
-
-    // Enable the MMU and turn on I-cache and D-cache
-    SCTLR_EL2.modify(SCTLR_EL2::M::Enable + SCTLR_EL2::C::Cacheable + SCTLR_EL2::I::Cacheable);
-    barrier::isb(barrier::SY);
-}
-
-#[cfg(feature = "hv")]
-unsafe fn cache_invalidate(cache_level: usize) {
-    core::arch::asm!(
-        r#"
-        msr csselr_el1, {0}
-        mrs x4, ccsidr_el1 // read cache size id.
-        and x1, x4, #0x7
-        add x1, x1, #0x4 // x1 = cache line size.
-        ldr x3, =0x7fff
-        and x2, x3, x4, lsr #13 // x2 = cache set number – 1.
-        ldr x3, =0x3ff
-        and x3, x3, x4, lsr #3 // x3 = cache associativity number – 1.
-        clz w4, w3 // x4 = way position in the cisw instruction.
-        mov x5, #0 // x5 = way counter way_loop.
-    // way_loop:
-    1:
-        mov x6, #0 // x6 = set counter set_loop.
-    // set_loop:
-    2:
-        lsl x7, x5, x4
-        orr x7, {0}, x7 // set way.
-        lsl x8, x6, x1
-        orr x7, x7, x8 // set set.
-        dc cisw, x7 // clean and invalidate cache line.
-        add x6, x6, #1 // increment set counter.
-        cmp x6, x2 // last set reached yet?
-        ble 2b // if not, iterate set_loop,
-        add x5, x5, #1 // else, next way.
-        cmp x5, x3 // last way reached yet?
-        ble 1b // if not, iterate way_loop
-        "#,
-        in(reg) cache_level,
-        options(nostack)
-    );
-}
-
-#[cfg(feature = "hv")]
-#[naked]
-#[no_mangle]
-#[link_section = ".text.boot"]
-unsafe extern "C" fn _start() -> ! {
-    // PC = 0x8_0000
-    // X0 = dtb
-    core::arch::asm!("
-        mov     x20, x0                 // save DTB pointer
-        // disable cache and MMU
-        mrs x1, sctlr_el2
-        bic x1, x1, #0xf
-        msr sctlr_el2, x1
-
-        // cache_invalidate(0): clear dl1$
-        mov x0, #0
-        bl  {cache_invalidate}
-        mov x0, #2
-        bl  {cache_invalidate}
-
-        mrs     x19, mpidr_el1
-        and     x19, x19, #0xffffff     // get current CPU id
-
-        adrp    x8, {boot_stack}        // setup boot stack
-        add     x8, x8, {boot_stack_size}
-        mov     sp, x8
-
-        bl      {switch_to_el2}         // switch to EL2
-        bl      {enable_fp}             // enable fp/neon
-        bl      {init_boot_page_table}
-        bl      {init_mmu_el2}
-
-        mov     x8, {phys_virt_offset}  // set SP to the high address
-        add     sp, sp, x8
-
-        mov     x0, x19                 // call rust_entry(cpu_id, dtb)
-        mov     x1, x20
-        ldr     x8, ={entry}
-        blr     x8
-        b      .",
-        cache_invalidate = sym cache_invalidate,
-        init_boot_page_table = sym init_boot_page_table,
-        init_mmu_el2 = sym init_mmu_el2,
-        switch_to_el2 = sym switch_to_el2,
-        enable_fp = sym enable_fp,
-        boot_stack = sym BOOT_STACK,
-        boot_stack_size = const TASK_STACK_SIZE,
-        phys_virt_offset = const axconfig::PHYS_VIRT_OFFSET,
-        entry = sym crate::platform::rust_entry,
-        options(noreturn),
-    );
 }
